Migration Tooling / Release Process (#27)
* use poetry

* add push to github docker registry

* fix release.yml formatting error

* add .dockerignore file

* remove docker/make files that are not used anymore

* split pypgstac code to migrate.py load.py to help organize, add migration path logic for multi-step incremental migrations

* add pydocstyle comments

* add new scripts to enable a staging process for sql migrations

* use scripts/stageversion to add steps between all versions

* use scripts/stageversion to add steps between all versions

* Add documentation on release process to README
