+++ pypgstac/pypgstac/__init__.py
@@ -1,1 +1,2 @@
+"""PyPGStac Version."""
 __version__ = "0.2.8"
