+++ scripts/bin/testdb
@@ -1,0 +1,50 @@
+#!/bin/bash
+
+set -e
+
+if [[ "${CI}" ]]; then
+    set -x
+fi
+
+function usage() {
+    echo -n \
+        "Usage: $(basename "$0")
+Runs tests for the project.
+
+This scripts is meant to be run inside the dev container.
+
+"
+}
+
+function bail() {
+    echo $1
+    exit 1
+}
+
+source $(dirname $0)/migra_funcs
+
+if [ "${BASH_SOURCE[0]}" = "${0}" ]; then
+    echo "Creating Temp DBs to check base and incremental migrations."
+    create_migra_dbs
+    trap drop_migra_dbs 0 2 3 15
+
+    echo "Setting Temp DB 1 to initial version 0.1.9"
+    pypgstac migrate --dsn $FROMDBURL --toversion 0.1.9
+    echo "Using incremental migrations to bring Temp DB 1 up to date"
+    pypgstac migrate --dsn $FROMDBURL
+    echo "Running PGTap Tests on Temp DB 1"
+    pgtap $FROMDBURL
+
+    echo "Using latest base migrations on Temp DB 2"
+    pypgstac migrate --dsn $TODBURL
+    echo "Running PGTap Tests on Temp DB 2"
+    pgtap $TODBURL
+
+    echo "Using Migra to test for any differences between incrementally migrated DB and DB created with latest base migration."
+    diffs=$(migra --schema pgstac --unsafe $FROMDBURL $TODBURL || echo "")
+    [ $(echo $diff | wc -l) -gt 1 ] && { echo "DIFFERENCES FOUND: $diff"; exit 1; }
+
+    echo "No differences found."
+
+
+fi
