+++ sql/999_version.sql
@@ -1,3 +1,1 @@
-SET SEARCH_PATH TO pgstac, public;
-
-INSERT INTO migrations (version) VALUES ('0.2.8');
+INSERT INTO pgstac.migrations (version) VALUES ('0.2.8');
